chore : delete BottleView and add BottleViewModel
